scripts/requirements.txt
faiss-cpu
sentence-transformers
yt-dlp
numpy
tqdm